Reorganising code
